Updating PHPUnit config to match version 8.5, splitting the testsuites based on directories
